+++ trung/banle.php
@@ -7,6 +7,15 @@ $danhmucnongtrai = new DanhMucNongTrai();
 $danhmucbanle = new DanhMucBanLe();
 $donggoi_list = $donggoi->get_all_list();
 $danhmucbanle_list = $danhmucbanle->get_all_list();
+if(isset($_POST['submit'])){
+    $banle_check = isset($_POST['banle_check']) ? $_POST['banle_check'] : '';
+    if($banle_check){
+        foreach ($banle_check as $key => $value) {
+            $check = isset($_POST['bl_'.$value]) ? $_POST['bl_'.$value] : 0;
+            $banle->id = $value; $banle->lock($check);
+        }
+    }
+}
 if($users->is_admin()){
     $banle_list =  $banle->get_all_list();
 } else {
@@ -32,10 +41,18 @@ if($users->is_admin()){
                 <h4 class="panel-title"><i class="fa fa-gears"></i> Thông tin Nơi bán lẻ</h4>
             </div>
             <div class="panel-body">
+                <?php if($users->is_admin()) : ?>
+                <form action="<?php echo $_SERVER['REQUEST_URI']; ?>" method="POST">
+                <button type="submit" name="submit" id="submit" value="OK" class="btn btn-success"><i class="fa fa-lock"></i> Cập nhật khóa dữ liệu</button>    
+                <?php endif; ?>
             	<a href="#modal-banle" data-toggle="modal" class="btn btn-primary m-10 thembanle"><i class="fa fa-plus"></i> Thêm mới</a>
+                <a href="../export_data.html?collect=banletrung&submit=OK" class="btn btn-primary"><i class="fa fa-file-excel-o"></i> Xuất Excel</a>
             	<table id="data-table" class="table table-striped table-bordered table-hovered">
             		<thead>
             			<tr>
+                            <?php if($users->is_admin()) : ?>
+                            <th >Khóa<input type="checkbox" name="check_all" id="check_all"></th>
+                            <?php endif; ?>
             				<th>STT</th>
             				<th>Tên sản phẩm</th>
             				<th>Nơi bán lẻ</th>
@@ -53,7 +70,13 @@ if($users->is_admin()){
             			$i=1;
             			foreach($banle_list as $bl){
                             $donggoi->id = $bl['id_donggoitrung'];$dg = $donggoi->get_one();
+                            $check_lock = isset($bl['lock']) ? $bl['lock'] : 0;
             				echo '<tr>';
+                            echo '<tr>';
+                            if($users->is_admin()) :
+                            echo '<input type="hidden" name="banle_check[]" value="'.$bl['_id'].'" />';                                
+                            echo '<td><input type="checkbox" value="1" name="bl_'.$bl['_id'].'" class="check" '.($check_lock == 1 ? ' checked' : '').'/></td>';
+                            endif;
             				echo '<td>'.$i.'</td>';
             				echo '<td>'.$dg['tensanpham'].'</td>';
                             echo '<td>';
@@ -74,14 +97,22 @@ if($users->is_admin()){
             					echo '<td class="text-center link_hienthi"><a href="get.banle.html?id='.$bl['_id'].'&hienthi=1&act=hienthi" class="sethienthi" onclick="return false;"><i class="fa fa-eye-slash text-danger"></i></a></td>';
             				}*/
             				echo '<td class="text-center"><a href="../print_qrcode_trung.html?id='.$bl['_id'].'&type=4&q=trung" class="open_window"><i class="fa fa-qrcode"></i></a></td>';
-            				echo '<td class="text-center"><a href="get.banle.html?id='.$bl['_id'].'&act=del" onclick="return confirm(\'Chắc chắn muốn xoá?\');"><i class="fa fa-trash"></i></a></td>';
-            				echo '<td class="text-center"><a href="get.banle.html?id='.$bl['_id'].'&act=edit#modal-banle" data-toggle="modal" class="suabanle"><i class="fa fa-pencil"></i></a></td>';
+                            if($check_lock == 1){
+                                echo '<td class="text-center"><i class="fa fa-lock text-danger"></i></td>';
+                                echo '<td class="text-center"><i class="fa fa-lock text-danger"></i></td>';
+                            } else {
+                				echo '<td class="text-center"><a href="get.banle.html?id='.$bl['_id'].'&act=del" onclick="return confirm(\'Chắc chắn muốn xoá?\');"><i class="fa fa-trash"></i></a></td>';
+                				echo '<td class="text-center"><a href="get.banle.html?id='.$bl['_id'].'&act=edit#modal-banle" data-toggle="modal" class="suabanle"><i class="fa fa-pencil"></i></a></td>';
+                            }
             				echo '</tr>'; $i++;
             			}
             		}
             		?>
             		</tbody>
             	</table>
+                <?php if($users->is_admin()) : ?>
+                </form>
+                <?php endif; ?>
             </div>
         </div>
     </div>
@@ -105,10 +136,12 @@ if($users->is_admin()){
                     	<?php
                     	if($donggoi_list){
                     		foreach($donggoi_list as $dg){
-                                $nongtrai->id = $dg['id_nongtraitrung'];$nt=$nongtrai->get_one();
-                                $danhmucnongtrai->id = $nt['id_dmnongtrai']; $dmnt = $danhmucnongtrai->get_one();
-                                $danhmucnhamay->id = $dg['id_dmnhamay']; $dmnm = $danhmucnhamay->get_one();
-                    			echo '<option value="'.$dg['_id'].'">'.$dmnm['ten'].' - '.$nt['madan'].' - '.$dg['tensanpham'].' - '.$dg['quicachdonggoi'].' - '.date("d/m/Y", $dg['ngaydonggoi']->sec).' - '.$dg['solo'].' - '.date("d/m/Y", $nt['ngaythuhoach']->sec).' - '.$dmnt['ten'].'</option>';
+                                if($users->is_admin() || $dg['id_congty'] == $id_congty){
+                                    $nongtrai->id = $dg['id_nongtraitrung'];$nt=$nongtrai->get_one();
+                                    $danhmucnongtrai->id = $nt['id_dmnongtrai']; $dmnt = $danhmucnongtrai->get_one();
+                                    $danhmucnhamay->id = $dg['id_dmnhamay']; $dmnm = $danhmucnhamay->get_one();
+                        			echo '<option value="'.$dg['_id'].'">'.$dmnm['ten'].' - '.$nt['madan'].' - '.$dg['tensanpham'].' - '.$dg['quicachdonggoi'].' - '.date("d/m/Y", $dg['ngaydonggoi']->sec).' - '.$dg['solo'].' - '.date("d/m/Y", $nt['ngaythuhoach']->sec).' - '.$dmnt['ten'].'</option>';
+                                }
                     		}
                     	}
                     	?>
@@ -192,6 +225,16 @@ if($users->is_admin()){
             time:""
         });
         <?php endif; ?>        
-        App.init();TableManageDefault.init();FormSliderSwitcher.init();
+        $("#check_all").click(function(){
+            if($(this).prop("checked")){
+                $(".check").prop("checked", true);
+            } else {   
+                $(".check").prop("checked", false);
+            }
+        });
+        App.init();FormSliderSwitcher.init();
+        <?php if(!$users->is_admin()): ?>
+            TableManageDefault.init();
+        <?php endif; ?>
     });
 </script>
